backend/requirements.txt
# ── FastAPI core ──────────────────────────────────────────────
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── Database ──────────────────────────────────────────────────
sqlalchemy==2.0.29
alembic==1.13.1
aiosqlite==0.20.0          # SQLite async (dev mode)
asyncpg==0.29.0            # PostgreSQL async (prod mode)
psycopg2-binary==2.9.9     # Sync fallback for Alembic

# ── Auth / Security ───────────────────────────────────────────
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
passlib[bcrypt]==1.7.4

# ── Network monitoring ────────────────────────────────────────
scapy==2.5.0

# ── Machine learning ──────────────────────────────────────────
scikit-learn==1.4.1.post1
numpy==1.26.4
joblib==1.3.2

# ── GeoIP ────────────────────────────────────────────────────
geoip2==4.7.0
maxminddb==2.5.2

# ── Scheduler ────────────────────────────────────────────────
APScheduler==3.10.4

# ── Docker SDK (Cowrie controller) ────────────────────────────
docker==7.0.0

# ── HTTP client (internal) ────────────────────────────────────
httpx==0.27.0

# ── Logging ───────────────────────────────────────────────────
structlog==24.1.0

# ── Validation / config ───────────────────────────────────────
pydantic==2.6.4
pydantic-settings==2.2.1

# ── WebSocket ────────────────────────────────────────────────
websockets==12.0

# ── Utilities ────────────────────────────────────────────────
python-dotenv==1.0.1
pytz==2024.1
